refactor ide-updater service
